Python/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
Django==3.2
django-cors-headers==3.7.0
django-filter==2.4.0
django-polymorphic==3.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
filelock==3.0.12
future==0.18.2
gevent==21.1.2
greenlet==1.0.0
idna==2.10
Markdown==3.3.4
psycopg2-binary==2.8.6
PyJWT==2.0.1
PySocks==1.7.1
pytz==2021.1
requests==2.25.1
requests-file==1.5.1
six==1.15.0
sqlparse==0.4.1
tldextract==3.1.0
urllib3==1.26.4
zope.event==4.5.0
zope.interface==5.4.0
